+++ src/viewport/viewport.const.ts
@@ -8,8 +8,10 @@ export const UX_VIEWPORT_DEFAULT_BREAKPOINTS: Dictionary<number> = {
 	medium: 992,
 	large: 1200,
 	xlarge: 1500,
-	xxlarge: 1920,
-	xxlarge1: 2100,
+	fhd: 1920,
+	qhd: 2560,
+	uhd4k: 3840,
+	uhd8k: 7680,
 };
 
 export const UX_VIEWPORT_DEFAULT_CONFIG: UxViewportOptions = {