+++ src/viewport/viewport.model.ts
@@ -10,8 +10,10 @@ export enum ViewportSizeType {
 	medium = 2,
 	large = 3,
 	xlarge = 4,
-	xxlarge = 5,
-	xxlarge1 = 6,
+	fhd = 5,
+	qhd = 6,
+	uhd4k = 7,
+	uhd8k = 8,
 }
 
 export enum ComparisonOperation {